+++ server/db/update_scripts/29_update_v6.14.0_v6.15.0.sql
@@ -2,5 +2,10 @@
 UPDATE version
 SET version = 'v6.15.0';
 
-ALTER TABLE scheduledJobs_formActions
-ADD COLUMN record_id INT(10) UNSIGNED ZEROFILL NOT NULL;
+--  for internal data type
+CALL add_table_column('scheduledJobs_formActions', 'id_uploadRows', 'INT(10) UNSIGNED ZEROFILL');
+CALL add_foreign_key('scheduledJobs_formActions', 'scheduledJobs_formActions_fk_id_uploadRows', 'id_uploadRows', 'uploadRows (id)');
+
+--  for external data type
+CALL add_table_column('scheduledJobs_formActions', 'id_user_input_record', 'INT(10) UNSIGNED ZEROFILL');
+CALL add_foreign_key('scheduledJobs_formActions', 'scheduledJobs_formActions_fk_id_user_input_record', 'id_user_input_record', 'user_input_record (id)');